elasticsearch/requirements.txt
numpy>=1.19.3
sentence_transformers>=0.4.1.2
elasticsearch==7.5.0
PyYAML>=5.4.1
